+++ src/main/kotlin/com/andredonadon/restaurantapi/entity/PlaceCategoryEntity.kt
@@ -2,10 +2,9 @@ package com.andredonadon.restaurantapi.entity
 
 import org.springframework.data.annotation.Id
 import org.springframework.data.relational.core.mapping.Table
-import java.math.BigInteger
 
 @Table("place_categories")
 data class PlaceCategoryEntity(
-    @Id val id: BigInteger? = null,
+    @Id val id: Int? = null,
     val description: String
 )
